requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==5.0.4
docutils==0.20.1
Flask==3.0.3
flask-cors==6.0.1
flask-restx==1.3.0
gunicorn==23.0.0
idna==3.4
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Kivy==2.2.1
Kivy-Garden==0.1.5
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==2.1.5
packaging @ file:///private/tmp/python-packaging-20231018-5918-1ntnn9h/packaging-23.2
pycairo @ file:///private/tmp/py3cairo-20231022-6959-nbqbp3/pycairo-1.25.1
pygame==2.6.1
Pygments==2.16.1
PyGObject==3.46.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-decouple==3.8
python-dotenv==1.1.0
pytube==15.0.0
pytz==2024.2
redis==6.2.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
soupsieve==2.6
spotipy==2.25.1
sqlparse==0.5.0
translate==3.6.1
urllib3==2.0.6
Werkzeug==3.0.4
